% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multilada_misc.R
\name{age_months}
\alias{age_months}
\title{Calculate age in months}
\usage{
age_months(dob, submission_date)
}
\arguments{
\item{dob, submission_date}{Vectors of dates in the "ymd" format.}
}
\value{
A \code{numeric} vector.
}
\description{
Properly (!) calculates age in months given date of birth and submission (testing etc.) date.
}
\details{
No argument testing yet implemented. They need to follow Ymd or ymd format strictly.
}
