% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdi_adaptive.R
\name{cdi_adaptive}
\alias{cdi_adaptive}
\title{Summarise adaptive CDI submissions}
\usage{
cdi_adaptive(data)
}
\arguments{
\item{data}{A \code{dataframe} containing answers to a \emph{CDI-Online} \strong{adaptive} form,
as returned by \code{\link[=cdi_read]{cdi_read()}} (or obtained otherwise).}
}
\value{
A \code{dataframe} summarising key information for each submission.
}
\description{
Get various information about adaptive CDI submissions.
}
\details{
So far it summarises submissions only.
Extracting information on individual responses, items, etc. might be added later
(here or in separate functions).
}
