% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdi.R
\name{cdi_count_checkboxAlt}
\alias{cdi_count_checkboxAlt}
\alias{cdi_count_manyCheckboxGroups}
\alias{cdi_count_radio}
\alias{cdi_count_oneCheckboxGroup}
\title{Count responses to given question type in CDI-Online form}
\usage{
cdi_count_checkboxAlt(data, type, category = NULL, answer = "first")

cdi_count_manyCheckboxGroups(data, type, category = NULL, answer = "first")

cdi_count_radio(data, type, category = NULL, answer = "1")

cdi_count_oneCheckboxGroup(data, type, category = NULL)
}
\arguments{
\item{data}{A \code{dataframe} containing answers to a \emph{CDI-Online} form,
e.g., as returned by \code{\link[=cdi_read]{cdi_read()}} (or obtained otherwise).}

\item{type}{A \code{character}, the value of the \code{type} column of the \code{data} dataframe,
e.g., "word" for \code{oneCheckboxGroup}, "alternatives" for \code{checkboxAlt}.}

\item{category}{A \code{character}, the value of the \code{category} column of the \code{data} dataframe,
e.g., "animals", "action_words", or "quantifiers" for the type "word" in \emph{CDI:WG} or \emph{CDI:WS}.
If set, only answers in this category are counted. Otherwise answers for the whole type are considered.}

\item{answer}{A \code{character} specifing which asnwers to count.
For \code{checkboxAlt} and \code{manyCheckboxGroups} possible values are
"first" (default), "second", "both", or "none"
(for \code{manyCheckboxGroups} "first" usually means \emph{understands} and "second" means \emph{produces}),
For \code{radio} it should be an integer
(coerced to \code{character}) indicating the option to count and defaulting to "1"
(which usually codes "yes" in \emph{yes or no} questions).}
}
\value{
A \code{dataframe} with columns: "id", "n" (resulting count), "type" (the argument to the function,
not \code{answer_type}), "answer" (if applicable), and "category" (if not \code{NULL}).
}
\description{
These functions count (given) responses to questions of a given type for each ID.
Types (aka \code{answer_type}s) implemented so far are:
\itemize{
\item \code{oneCheckboxGroup}: type used for simple checkbox lists (e.g., word lists in \emph{CDI:WS}, \emph{CDI-III});
\item \code{manyCheckboxGroups}: type used for checkbox lists with two options (e.g., word list in \emph{CDI:WG});
\item \code{checkboxAlt}: type used for checkbox alternatives (e.g., some \emph{Complexity} sections);
\item \code{radio}: traditional radio button question type.
}
}
\details{
To find out what \code{type} or \code{category} you are looking for, check the \code{data} dataframe or
\code{items.csv} for a given form in the \verb{www/languages/} of your \emph{CDI-Online} installation.
To find out which function to use, check the \code{answer_type} column in the \code{data} dataframe or
the \code{type} column in \code{settings.csv} for a given form.

\code{type} is the highest level in the hierarchy of items in \emph{CDI-Online} forms.
\code{category} is a sub-\code{type} and also a basic level of the item hierarchy,
displayed as a single page. Hence sometimes \code{categories} are meaningful
but sometimes they are used only for the sake of the UX:
they divide longer \code{types} into a number of pages, e.g., "complexity1", "complexity2" etc.
(as a matter of fact \code{type} can also be used just for that reason,
e.g., "gestures_first", "gestures_games" etc. without overarching "gestures").
}
\examples{
\dontrun{
cdi_count_oneCheckboxGroup(ws_pl, "word")
cdi_count_oneCheckboxGroup(ws_pl, "word", "vehicles")
cdi_count_checkboxAlt("cdi3-1_pl", "alternatives")
}

}
