% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdi_connect.R
\name{cdi_forms}
\alias{cdi_forms}
\alias{cdi_read}
\title{Connect to a CDI-Online database}
\usage{
cdi_forms(database = NULL)

cdi_read(database = NULL, form)
}
\arguments{
\item{database}{A \code{character} identifying a Multilada database or, to be precise,
all secret credentials needed to connect to the database (e.g., "host", "port").
If \code{database = NULL} (the default), you are asked to provide each credential when prompted.
Otherwise the function attempts to retrieve the credentials for the database from
the system-based credential store (e.g., \emph{Keychain} on macOS). If it fails to find
a given credential, you are asked to provide it and then
it is saved in the credential store for future use.}

\item{form}{A \code{character}, name of the table to read. It's best to run cdi_forms() first
to see available options.}
}
\value{
For \code{cdi_forms()} a \code{character} vector of table names for \emph{CDI-Online} forms.
\emph{CDI-Online} app creates a name for a table by appending the value of parameter \code{lang}
to the value of parameter \code{form} using "_" as a separator.

For \code{cdi_read()} a data frame with all responses for a given form. The format of the responses mirrors
the way \emph{CDI-Online} app saves them to a table in a database.
}
\description{
\code{cdi_forms()} gets a list of all \emph{CDI-Online} forms stored as tables
with saved responses in a given database.
\code{cdi_read()} retrieves all responses to a given form (a table in a database).
}
\details{
The underlying workhorse for connecting to a database is \code{\link[=multilada_connect]{multilada_connect()}}
and ultimately \code{\link[=multilada_credentials]{multilada_credentials()}} for dealing with database credentials.
Check the latter to better understand the credentials asking/saving/retrieving details.
}
\examples{
# To print list of names of all CDI forms stored in a given database:
\dontrun{
cdi_forms("cdi")}
# If you have already saved the credentials needed to connect
# to that database under the name "cdi" in your system,
# they will be retrevied. Otherwise you will be asked for them
# and they will be saved under the name "cdi" for future use.
# ---
#
# If there is a form saved under the name "ws_pl" in that database,
# to import it too a dataframe:
\dontrun{
ws_pl <- cdi_read("cdi", "ws_pl")}

}
