% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdi_itemise.R
\name{cdi_itemise_checkboxAlt}
\alias{cdi_itemise_checkboxAlt}
\title{Add item definitions to \code{checkboxAlt} data}
\usage{
cdi_itemise_checkboxAlt(data, type, items)
}
\arguments{
\item{data}{A data frame containing answers to a \emph{CDI-Online} \strong{static} form,
as returned by \code{\link[=cdi_read]{cdi_read()}} (or obtained otherwise).}

\item{type}{A \code{character}, the value of the \code{type} column of the \code{data} dataframe,
e.g., "word" for \code{oneCheckboxGroup}, "alternatives" for \code{checkboxAlt}.}

\item{items}{A data frame containing item definitions for a \emph{CDI-Online} form, e.g., as imported from
\code{items.csv} file found in the appropriate subfolder of \verb{www/languages/} in your \emph{CDI-Online} installation
(required columns are \code{type}, \code{category}, and \code{definition}).}
}
\value{
A data frame with columns: "id", "type", "category", "item_id" (sequence number within category),
"answer1" (0 or 1), "answer2" (0 or 1), and "definition", as well as any other columns contained in \code{items}.
}
\description{
Add item definitions to \code{checkboxAlt} data
}
