% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdi_itemise.R
\name{cdi_itemise_oneCheckboxGroup}
\alias{cdi_itemise_oneCheckboxGroup}
\title{Add item definitions to \code{oneCheckboxGroup} data}
\usage{
cdi_itemise_oneCheckboxGroup(data, items)
}
\arguments{
\item{data}{A \code{dataframe} containing answers to a \emph{CDI-Online} \strong{static} form,
as returned by \code{\link[=cdi_read]{cdi_read()}} (or obtained otherwise).}

\item{items}{A \code{dataframe} containing item definitions for a \emph{CDI-Online} form, e.g., as imported from
\code{items.csv} file found in the appropriate subfolder of \verb{www/languages/} in your \emph{CDI-Online} installation
(required columns are \code{type}, \code{category}, and \code{definition}).}
}
\value{
A \code{dataframe} with columns: "id", "type", "category", "item_id" (sequence number within category),
"response" (0 or 1), and "definition", as well as any other columns contained in \code{items}.
}
\description{
Results for \code{oneCheckboxGroup} items (e.g., word lists in \emph{CDI:WS}, \emph{CDI-III})
contain only checked items with their sequence numbers within a category, i.e., a single page
without the rest of items and item definitions.
\code{cdi_itemise_oneCheckboxGroup()} adds these missing information,
which is helpful for item-level analyses.
}
