% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdi.R
\name{cdi_submissions}
\alias{cdi_submissions}
\alias{cdi_time}
\title{List IDs of all submissions together with start dates, end dates, and filling times}
\usage{
cdi_submissions(data, run = FALSE, form = FALSE, sort.by.end = FALSE)

cdi_time(data, run = FALSE, form = FALSE, sort.by.end = FALSE)
}
\arguments{
\item{data}{A \code{dataframe} containing answers to a \emph{CDI-Online} form,
e.g., as returned by \code{\link[=cdi_read]{cdi_read()}} (or obtained otherwise).}

\item{run}{A logical value indicating whether to include column "run".}

\item{form}{A logical value indicating whether to include column "form".}

\item{sort.by.end}{A logical value indicating whether to sort by end date
(by default the resulting \code{tibble} is sorted by start date).}
}
\value{
A \code{tibble} with columns: "id", "start" ("start_date" in the \code{data} dataframe),
"end" ("end_date" in the \code{data} dataframe), "duration"
(difference between "end_date" and "start_date" in the \code{data} dataframe), and
optionally "run".
}
\description{
Useful to get a summary of submissions.
\code{cdi_time()} is an alias for backward compatibility.
}
