% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fm_read.R
\name{fm_fields}
\alias{fm_fields}
\title{Prepare WP \emph{Form Maker} field names}
\usage{
fm_fields(source_file, lang)
}
\arguments{
\item{source_file}{A \code{character}, the name of a \code{csv} file downloaded
using the "Export to CSV" option of the WP \emph{Form Maker} plugin}

\item{lang}{A \code{character}, the name of resulting column, e.g.,
a language ISO code ("no", "pl" etc.).}
}
\value{
A single column tibble with \code{lang} as the name of the column.
}
\description{
Reads column names of a \code{csv} file downloaded using the "Export to CSV" option
of the WordPress \emph{Form Maker} plugin. "Repairs" duplicated names, as
required by \code{\link[=fm_read]{fm_read()}}, and returns a single column tibble, which
can be used to create a column with field names in a given language
in \code{variables_file} required by \code{\link[=fm_read]{fm_read()}}.
}
\details{
It basically applies \code{\link[vctrs:vec_as_names]{vctrs::vec_as_names()}} with \code{repair = "unique"},
adding "...j" to each duplicated column, where \code{j} is the column position
(consistent with \code{\link[readr:read_delim]{readr::read_csv()}}'s default behaviour used in \code{\link[=fm_read]{fm_read()}}.
}
\seealso{
\code{\link[=fm_read]{fm_read()}}
}
