% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fm_process.R
\name{fm_matrix}
\alias{fm_matrix}
\alias{fm_grading}
\title{Preprocess responses to \emph{WP Form Maker} matrix-type questions}
\usage{
fm_matrix(data, names_prefix = NULL, keep = FALSE)

fm_grading(data, names_prefix = NULL, keep = FALSE)
}
\arguments{
\item{data}{blah}

\item{names_prefix}{blah}

\item{keep}{blah}
}
\value{
description
}
\description{
Matrix questions...
}
\details{
Additional details...
}
\examples{
# example code

}
