% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fm_read.R
\name{fm_read}
\alias{fm_read}
\title{Read and preprocess \emph{WP Form Maker} data}
\usage{
fm_read(data_file, variables_file, translations_file = NULL, lang)
}
\arguments{
\item{data_file}{description}

\item{variables_file}{description}

\item{translations_file}{description}

\item{lang}{description}
}
\value{
description
}
\description{
Reads a \code{csv} file downloaded using the "Export to CSV" option
of the \emph{WP Form Maker} plugin.
}
\details{
Additional details...
}
\examples{
# example code

}
