% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fm_read.R
\name{fm_variables}
\alias{fm_variables}
\title{Prepare \code{variables_file} for \code{fm_read()}}
\usage{
fm_variables(
  source_file,
  lang,
  simplified = FALSE,
  target_file = NULL,
  csv2 = FALSE
)
}
\arguments{
\item{source_file}{A \code{character}, the name of a \code{csv} file downloaded
using the "Export to CSV" option of the WP \emph{Form Maker} plugin.}

\item{lang}{A \code{character}, the name of the column with field names, e.g.,
a language ISO code ("no", "pl" etc.).}

\item{simplified}{A logical value. If set to \code{FALSE} (default), the resulting tibble
contains all the columns required by \code{\link[=fm_read]{fm_read()}}
("Import", "Label", and "Type" filled with \code{NA}s).
If set to \code{TRUE}, a single column tibble is returned
(only column \code{lang} with field names).}

\item{target_file}{A \code{character}, an optional name of file to which
the resulting tibble will be saved.}

\item{csv2}{A logical value. If set to \code{FALSE} (default), the source file is read
using \code{readr::read_csv()}. If set to \code{TRUE}, \code{readr::read_csv2()} is used.}
}
\value{
A tibble with \code{lang} and optionally "Import", "Label", and "Type" columns.
If \code{target_file} is specified, returns invisibly.
}
\description{
Reads column names of a \code{csv} file downloaded using the "Export to CSV" option
of the WordPress \emph{Form Maker} plugin (i.e. it reads field names of a given form).
"Repairs" duplicated names, as required by \code{\link[=fm_read]{fm_read()}}, returns a tibble, which
can be used to create \code{variables_file} required by \code{\link[=fm_read]{fm_read()}}, and
optionally saves it as such a file.
}
\details{
For field names (column names of the \emph{FM} responses file)
it basically applies \code{\link[vctrs:vec_as_names]{vctrs::vec_as_names()}} with \code{repair = "unique"},
adding "...j" to each duplicated column, where \code{j} is the column position
(consistent with \code{\link[readr:read_delim]{readr::read_csv()}}'s default behaviour used in \code{\link[=fm_read]{fm_read()}}.
The option \code{csv2 = TRUE} adds the support for reading a \code{source_file} with
a non-comma separator (e.g., a semicolon).
}
\seealso{
\code{\link[=fm_read]{fm_read()}}
}
