% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/credentials.R
\name{multilada_credentials_element}
\alias{multilada_credentials_element}
\title{Get or set credentials securely}
\usage{
multilada_credentials_element(element, key, prompt)
}
\description{
The workhorse of \code{multilada_credentials()}.
}
