% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/credentials.R
\name{multilada_credentials_file_get}
\alias{multilada_credentials_file_get}
\title{Get credentials from encrypted file}
\usage{
multilada_credentials_file_get(
  key,
  elements = c("host", "port", "name", "username", "password")
)
}
\description{
\strong{File backend for \code{keyring} doesn't work on Windows}.
}
