% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/credentials.R
\name{multilada_credentials_file_where}
\alias{multilada_credentials_file_where}
\title{Check encrypted credentials file's location}
\usage{
multilada_credentials_file_where(key = NULL)
}
\description{
\strong{File backend for \code{keyring} doesn't work on Windows}.
}
