% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdi_submissions.R
\name{cdi_submissions}
\alias{cdi_submissions}
\alias{cdi_time}
\title{Summarise all submissions}
\usage{
cdi_submissions(data, run = FALSE, form = FALSE, sort.by.end = FALSE)

cdi_time(data, run = FALSE, form = FALSE, sort.by.end = FALSE)
}
\arguments{
\item{data}{A \code{dataframe} containing answers to a \emph{CDI-Online} form,
e.g., as returned by \code{\link[=cdi_read]{cdi_read()}} (or obtained otherwise).}

\item{run}{A logical value indicating whether to include column "run".}

\item{form}{A logical value indicating whether to include column "form".}

\item{sort.by.end}{A logical value indicating whether to sort by end date
(by default the resulting \code{tibble} is sorted by start date).}
}
\value{
A \code{tibble} with the following columns:
\itemize{
\item "id",
\item optionally "run" and "form",
\item "start" ("start_date" in the \code{data} dataframe),
"end" ("end_date" in the \code{data} dataframe), "duration"
(difference between "end_date" and "start_date" in the \code{data} dataframe),
\item "birth_date", "sex", "guardian" (\emph{What is your relationship to the child?}),
and possibly "guardian.other", if relevant data collected for a given form
("demographic" input type in \code{settings.csv}).
}
}
\description{
Useful to get a list of IDs together with basic stats.
\code{cdi_time()} is an alias for backward compatibility.
}
