% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fm_process.R
\name{fm_matrix}
\alias{fm_matrix}
\alias{fm_grading}
\title{Parse responses to \emph{WP Form Maker} multiple item fields}
\usage{
fm_matrix(data, names_prefix = NULL, keep = FALSE)

fm_grading(data, names_prefix = NULL, keep = FALSE)
}
\arguments{
\item{data}{A data frame with at least two columns.
The first column is expected to contain IDs and the second column
should be the column to be parsed. NB this a very "non-tidyverse" and
very base \emph{R} approach but it keeps the parsed responses linked
to the appropriate IDs and allows limiting parsing to a subset of IDs
(if you're fluent in base \emph{R}...).}

\item{names_prefix}{An optional \code{character} providing a prefix to the names of
resulting columns. If \code{NULL} (the default) the names are constructed by prefixing
each item label with the name of the original column (with "_" as
a connector). Use this opion if you want to change this prefix
(perhaps to something simpler).}

\item{keep}{\code{Logical}. Should the original column be kept in the returned
data frame. Defaults to \code{FALSE}. Setting to \code{TRUE} is helpful when you want
to check if the parsing works as expected. To merge the parsed responses back
to the original \emph{Form Maker} data frame you should keep the default
(the original column is already there).}
}
\value{
A data frame with the first (ID) column of \code{data}, the resulting
columns with responses to each item, and optionally with the second
(parsed) column of \code{data}.
}
\description{
\code{fm_matrix()} parses responses to (radio) "matrix" fields:
same scale applied to a number of items.
\code{fm_grading()} parses responses to (numeric) so-called "grading" fields:
same type of response to a number of items.
\code{csv} files downloaded with the "Export to CSV" option of the WP \emph{Form Maker} plugin
keep all such responses in a single column, making it impossible to analyse them.
}
\details{
"Grading" fields should accept only numeric input but for some reason
(a bug in \emph{Form Maker}?) it hasn't always behaved like this and \code{fm_grading()}
accepts \code{character} entries as well.
}
\examples{
\dontrun{
# To check the parsing first:
fm_grading(df[, c("id", "first_contact_lang")], keep = TRUE) \%>\% view()

# To add the parsed responses back to the original df:
left_join(df, fm_grading(df[, c("id", "first_contact_lang")])) -> df
}

}
